dpe: check if CLARA_HOME env variable is set
If not set, the static initialization of the clara_home string variable
was throwing an exception or a segmentation fault.

Replace the static variable with a field and throw an exception with a
proper message. Using a field allows checking the environment variable
only when the DPE is used, instead of every time the library is loaded.