service: reduce the time the thread-pool workers wait for tasks
The thread-pool workers sleep for 1 millisecond if there are no tasks to
be executed. This adds a noticeable delay when the services are single
thread and their execution time is in the order of a few milliseconds.

For now make the worker sleep microseconds instead, to reduce the time
the tasks are waiting in the queue. The problem is an increased CPU
usage of the workers while checking for new tasks.