thread_pool: use a blocking queue
Use moodycamel::BlockingConcurrentQueue [1] to get tasks as soon as
possible instead of sleeping a fixed amount of time before checking for
new tasks.

This adds a proper fix to the workaround introduced by 5fff56a3bde1
("service: reduce the time the thread-pool workers wait for tasks").
Even with the reduced time, the waiting time added to the average event
time was still around half a ms for a simple chain of 3 services on
single thread, which may be too much when sending a lot of events (and
also the extra CPU usage was not ideal).

[1]: https://github.com/cameron314/concurrentqueue