clara-msg: update timeout of subscription tests
The sync-publish tests are failing on macOS due to the timeout being too
low (they are taking way longer than on Linux).

Increase the timeout values to 20000 (single producer) and to 10000
(multiple producers) to make them pass. Also, increase the timeout for
the subscription tests to 5000 just to give them a bit more of time (but
those are not failing).