message: remove arrays of primitives from default types
Remove support for default "array of primitives" data types. There is no
good reason to provide arrays as part of the default supported types, it
is a very specific case, and most probably a service using arrays will
have its own data type and serialization for it.

Also, they are stored in an ugly custom `Data` Protobuf message along
with the default supported primitive types, which I want to get rid of.